ExplicitImports

Documentation for ExplicitImports.

ExplicitImports.analyze_all_names

ExplicitImports.explicit_imports

ExplicitImports.find_implicit_imports

ExplicitImports.get_names_used

ExplicitImports.analyze_all_names — Method

analyze_all_names(file) -> DataFrame

Returns a DataFrame with one row per name per scope, with information about whether or not it is within global scope, what modules it is in, and whether or not it was assigned before ever being used in that scope.

ExplicitImports.explicit_imports — Function

explicit_imports(mod, file=pathof(mod); skips=(Base, Core)) -> Vector{String}

Returns a list of explicit import statements one could make for the module mod.

Notes:

Currently, this does not filter to only new explicit imports (these may be redundant with already existing explicit imports). That is, mod may already be explicitly importing these names.

Currently, this suggests explicit imports for names available through implicit exports even if every usage of the name is qualified (e.g. DataFrames.DataFrame).

if mod is not from a package, pathof will be unable to find the code, and a file must be passed which contains mod (either directly or indirectly through includes)

mod can be a submodule defined within file, but if two modules have the same name (e.g. X.Y.X and X), results may be inaccurate.

skips will skip names coming from the listed modules, or any submodules thereof.

ExplicitImports.find_implicit_imports — Method

find_implicit_imports(mod::Module; skips=(Base, Core))

Given a module mod, returns a Dict{Symbol, Module} showing names exist in mod's namespace which are available due to implicit exports by other modules. The dict's keys are those names, and the values are the module that the name comes from.

In the case of ambiguities (two modules exporting the same name), the name is unavailable in the module, and hence the name will not be present in the dict.

This is powered by Base.which.

ExplicitImports.get_names_used — Method

get_names_used(file) -> DataFrame

Figures out which global names are used in file, and what modules they are used within.

Traverses static include statements.

Returns a DataFrame with two columns, one for the name, and one for the path of modules, where the first module in the path is the innermost.