API

The main entrypoint for interactive use is print_explicit_imports. ExplicitImports.jl API also includes several other functions to provide programmatic access to the information gathered by the package, as well as utilities to use in regression testing.

ExplicitImports.print_explicit_imports — Function

print_explicit_imports([io::IO=stdout,] mod, file=pathof(mod); kw...)

Runs explicit_imports and prints the results, along with those of stale_explicit_imports. Accepts the same keyword arguments as that function.

ExplicitImports.explicit_imports — Function

explicit_imports(mod, file=pathof(mod); skips=(Base, Core), warn=true)

Returns a nested structure providing information about explicit import statements one could make for each submodule of mod.

file=pathof(mod): this should be a path to the source code that contains the module mod.

if mod is not from a package, pathof will be unable to find the code, and a file must be passed which contains mod (either directly or indirectly through includes)

mod can be a submodule defined within file, but if two modules have the same name (e.g. X.Y.X and X), results may be inaccurate.

skips=(Base, Core): any names coming from the listed modules (or any submodules thereof) will be skipped.

warn=true: whether or not to warn about stale explicit imports.

See also print_explicit_imports to easily compute and print these results, and explicit_imports_single for a non-recursive version which ignores submodules.

ExplicitImports.stale_explicit_imports — Function

stale_explicit_imports(mod, file=pathof(mod)) -> Vector{Symbol}

Returns a list of names that are not used in mod, but are still explicitly imported.

ExplicitImports.explicit_imports_single — Function

explicit_imports_single(mod, file=pathof(mod); skips=(Base, Core), warn=true)

A non-recursive version of explicit_imports; see that function for details.

Usage in testing

ExplicitImports.jl provides two functions which can be used to regression test that there is no reliance on implicit imports or stale explicit imports:

ExplicitImports.check_no_implicit_imports — Function

check_no_implicit_imports(mod, file=pathof(mod); skips=(Base, Core), warn=false)

Checks that neither mod nor any of its submodules is relying on implicit imports, throwing an ImplicitImportsException if so, and returning nothing otherwise.

This can be used in a package's tests, e.g.

@test check_no_implicit_imports(MyPackage) === nothing

ExplicitImports.check_no_stale_explicit_imports — Function

check_no_stale_explicit_imports(mod, file=pathof(mod))

Checks that neither mod nor any of its submodules has stale (unused) explicit imports, throwing an StaleImportsException if so, and returning nothing otherwise.

This can be used in a package's tests, e.g.

@test check_no_stale_explicit_imports(MyPackage) === nothing